snapshotstate: move sizer to osutil.Sizer()
This commit moves the "sizer" io.Writer out into the osutil package
because it turns out we need a writer that just counts the amount
of bytes written in various places so it makes sense to have a
helper for this.
